Add Nikita Nagorny's bachelor thesis as an example
The LaTeX source was taken from Nikita's Github project [1], imported into
LyX (using File->Import feature), and configured to be build with this
repository's physvsuThesis.cls file as the document class.

Currently, it can be built but the example itself is not finished in some
details: the typesetting is not perfect yet, annotiations are to be written
and extended.

[1]: https://github.com/N-Nagorny/bachelor-diploma

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>